double test

+++ app/src/main/java/com/example/mini_jeu_calcul/SystemCalcul.java
@@ -138,7 +138,7 @@ public class SystemCalcul {
         CalculerResultat(Taille);
     }
 
-public boolean TestReponse (int Reponse)
+public boolean TestReponse (double Reponse)
 {
         if (Reponse == Resultat){
             return true;
